Cleanup of Datasets

+++ rust/src/tasks/multi_label/mod.rs
@@ -14,11 +14,11 @@ trait ConfigTypes {
     type DataType;
 }
 
-#[derive(Clone, Serialize, Deserialize)]
+#[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct MultiConfig{
-    pub batch_size:u32,
-    pub sequence_length:u32,
-    pub number_labels:u32,
+    pub batch_size:usize,
+    pub sequence_length:usize,
+    pub number_labels:usize,
     pub tokenizer_name:String
 }
